# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Xendit Rides API",
    "description": "This is the Xendit Backend Exam for Software Engineer position. Below, you can find the API endpoints you can use. You can find out more about Swagger at http://swagger.io.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8010",
  "basePath": "",
  "tags": [
    {
      "name": "Rides",
      "description": "API for Xendit rides"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Rides"
        ],
        "summary": "Check health",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/rides": {
      "get": {
        "tags": [
          "Rides"
        ],
        "summary": "Get rides",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "description": "Desired page to fetch",
            "type": "integer"
          },
          {
            "name": "recordsPerPage",
            "in": "query",
            "description": "Number of records per page",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Rides"
            }
          },
          "404": {
            "description": "No rides found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "tags": [
          "Rides"
        ],
        "summary": "Create a new ride",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Ride that will be created",
            "schema": {
              "$ref": "#/definitions/Ride"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Ride"
            }
          },
          "400": {
            "description": "Invalid data supplied"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/rides/{id}": {
      "get": {
        "tags": [
          "Rides"
        ],
        "summary": "Find ride by ID",
        "description": "Returns a single ride",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the ride",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Ride"
            }
          },
          "404": {
            "description": "Ride not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "definitions": {
    "Ride": {
      "type": "object",
      "properties": {
        "rideID": {
          "type": "integer"
        },
        "start_lat": {
          "type": "integer"
        },
        "start_long": {
          "type": "integer"
        },
        "end_lat": {
          "type": "integer"
        },
        "end_long": {
          "type": "integer"
        },
        "rider_name": {
          "type": "string"
        },
        "driver_name": {
          "type": "string"
        },
        "driver_vehicle": {
          "type": "string"
        },
        "created": {
          "type": "string"
        }
      }
    },
    "Rides": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/Ride"
          }
        },
        "links": {
          "type": "object",
          "properties": {
            "prev_page": {
              "type": "string"
            },
            "next_page": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}